+++ common/overlays.nix
@@ -19,8 +19,6 @@ in
 
   linux = [
     self.overlays.element-desktop
-    self.overlays.fcitx-engines
-    self.overlays.jellyfin-wayland
     self.overlays.makeModulesClosure
     self.overlays.moonlight-wayland
     self.overlays.mpvpaper